fix schemas
